+++ src/Projectile.cpp
@@ -16,7 +16,7 @@ bool Projectile::reachedDestination() const {
 
 sf::Sprite Projectile::drawable() {
 	proj.move(m_offset);
-	proj.rotate(m_spin);
+	proj.rotate(static_cast<float>(m_spin));
 	return proj;
 }
